로그인 검증 기능 계층 분리
