+++ app/src/main/java/com/skysmyoo/publictalk/data/source/remote/UserRemoteDataSource.kt
@@ -1,11 +1,19 @@
 package com.skysmyoo.publictalk.data.source.remote
 
 import android.net.Uri
+import com.google.firebase.database.DataSnapshot
+import com.google.firebase.database.DatabaseError
+import com.google.firebase.database.ValueEventListener
+import com.google.firebase.database.ktx.database
+import com.google.firebase.ktx.Firebase
+import com.skysmyoo.publictalk.BuildConfig
 import com.skysmyoo.publictalk.data.model.remote.User
 import com.skysmyoo.publictalk.utils.TimeUtil
 import kotlinx.coroutines.tasks.await
 import retrofit2.Response
 import javax.inject.Inject
+import kotlin.coroutines.resume
+import kotlin.coroutines.suspendCoroutine
 
 class UserRemoteDataSource @Inject constructor(private val apiClient: ApiClient) {
     suspend fun putUser(auth: String, user: User): Response<Map<String, String>> {
@@ -30,6 +38,22 @@ class UserRemoteDataSource @Inject constructor(private val apiClient: ApiClient)
             downloadUri.toString()
         } else {
             null
+        }
+    }
+
+    suspend fun isExistUser(email: String?): DataSnapshot? {
+        val ref = Firebase.database(BuildConfig.BASE_URL).getReference("users")
+        return suspendCoroutine { continuation ->
+            ref.orderByChild("userEmail").equalTo(email)
+                .addListenerForSingleValueEvent(object : ValueEventListener {
+                    override fun onDataChange(snapshot: DataSnapshot) {
+                        continuation.resume(snapshot)
+                    }
+
+                    override fun onCancelled(error: DatabaseError) {
+                        continuation.resume(null)
+                    }
+                })
         }
     }
 }
